+++ Sources/ZcashLightClientKit/Block/Processor/CompactBlockScanning.swift
@@ -9,88 +9,83 @@
 import Foundation
 
 extension CompactBlockProcessor {
-    func compactBlockBatchScanning(range: CompactBlockRange) async throws {
-        try Task.checkCancellation()
-        
-        state = .scanning
 
-        // TODO: remove this arbitrary batch size https://github.com/zcash/ZcashLightClientKit/issues/576
-        let batchSize = scanBatchSize(for: range, network: self.config.network.networkType)
-        
+    func scanBlocks(at range: CompactBlockRange, totalProgressRange: CompactBlockRange) async throws {
+        try await compactBlockBatchScanning(range: range, totalProgressRange: totalProgressRange) { [weak self] lastScannedHeight in
+            let progress = BlockProgress(
+                startHeight: totalProgressRange.lowerBound,
+                targetHeight: totalProgressRange.upperBound,
+                progressHeight: lastScannedHeight
+            )
+            await self?.notifyProgress(.syncing(progress))
+        }
+    }
+
+    func compactBlockBatchScanning(
+        range: CompactBlockRange,
+        totalProgressRange: CompactBlockRange,
+        didScan: ((BlockHeight) async -> Void)? = nil
+    ) async throws {
+        try Task.checkCancellation()
+
         do {
-            if batchSize == 0 {
+            let scanStartHeight = try transactionRepository.lastScannedHeight()
+            let targetScanHeight = range.upperBound
+
+            var scannedNewBlocks = false
+            var lastScannedHeight = scanStartHeight
+
+            repeat {
+                try Task.checkCancellation()
+
+                let previousScannedHeight = lastScannedHeight
+
+                // TODO: remove this arbitrary batch size https://github.com/zcash/ZcashLightClientKit/issues/576
+                let batchSize = scanBatchSize(startScanHeight: previousScannedHeight + 1, network: self.config.network.networkType)
+
                 let scanStartTime = Date()
-                guard self.rustBackend.scanBlocks(dbCache: config.cacheDb, dbData: config.dataDb, limit: batchSize, networkType: config.network.networkType) else {
+                guard self.rustBackend.scanBlocks(
+                    dbCache: config.cacheDb,
+                    dbData: config.dataDb,
+                    limit: batchSize,
+                    networkType: config.network.networkType
+                ) else {
                     let error: Error = rustBackend.lastError() ?? CompactBlockProcessorError.unknown
                     LoggerProxy.debug("block scanning failed with error: \(String(describing: error))")
                     throw error
                 }
                 let scanFinishTime = Date()
-                
-                SDKMetrics.shared.pushProgressReport(
-                    progress: BlockProgress(
-                        startHeight: range.lowerBound,
-                        targetHeight: range.upperBound,
-                        progressHeight: range.upperBound
-                    ),
-                    start: scanStartTime,
-                    end: scanFinishTime,
-                    batchSize: Int(batchSize),
-                    operation: .scanBlocks
-                )
-                
-                let seconds = scanFinishTime.timeIntervalSinceReferenceDate - scanStartTime.timeIntervalSinceReferenceDate
-                LoggerProxy.debug("Scanned \(range.count) blocks in \(seconds) seconds")
-            } else {
-                let scanStartHeight = try transactionRepository.lastScannedHeight()
-                let targetScanHeight = range.upperBound
-                
-                var scannedNewBlocks = false
-                var lastScannedHeight = scanStartHeight
-                
-                repeat {
-                    try Task.checkCancellation()
-                    
-                    let previousScannedHeight = lastScannedHeight
-                    let scanStartTime = Date()
-                    guard self.rustBackend.scanBlocks(
-                        dbCache: config.cacheDb,
-                        dbData: config.dataDb,
-                        limit: batchSize,
-                        networkType: config.network.networkType
-                    ) else {
-                        let error: Error = rustBackend.lastError() ?? CompactBlockProcessorError.unknown
-                        LoggerProxy.debug("block scanning failed with error: \(String(describing: error))")
-                        throw error
-                    }
-                    let scanFinishTime = Date()
-                    
-                    lastScannedHeight = try transactionRepository.lastScannedHeight()
-                    
-                    scannedNewBlocks = previousScannedHeight != lastScannedHeight
-                    if scannedNewBlocks {
-                        let progress = BlockProgress(startHeight: scanStartHeight, targetHeight: targetScanHeight, progressHeight: lastScannedHeight)
-                        notifyProgress(.scan(progress))
-                        
-                        SDKMetrics.shared.pushProgressReport(
-                            progress: progress,
-                            start: scanStartTime,
-                            end: scanFinishTime,
-                            batchSize: Int(batchSize),
-                            operation: .scanBlocks
-                        )
-                        
-                        let heightCount = lastScannedHeight - previousScannedHeight
-                        let seconds = scanFinishTime.timeIntervalSinceReferenceDate - scanStartTime.timeIntervalSinceReferenceDate
-                        LoggerProxy.debug("Scanned \(heightCount) blocks in \(seconds) seconds")
-                    }
-                    
-                    await Task.yield()
-                } while !Task.isCancelled && scannedNewBlocks && lastScannedHeight < targetScanHeight
-                if Task.isCancelled {
-                    state = .stopped
-                    LoggerProxy.debug("Warning: compactBlockBatchScanning cancelled")
+
+                lastScannedHeight = try transactionRepository.lastScannedHeight()
+
+                scannedNewBlocks = previousScannedHeight != lastScannedHeight
+                if scannedNewBlocks {
+                    await didScan?(lastScannedHeight)
+
+                    let progress = BlockProgress(
+                        startHeight: totalProgressRange.lowerBound,
+                        targetHeight: totalProgressRange.upperBound,
+                        progressHeight: lastScannedHeight
+                    )
+
+                    SDKMetrics.shared.pushProgressReport(
+                        progress: progress,
+                        start: scanStartTime,
+                        end: scanFinishTime,
+                        batchSize: Int(batchSize),
+                        operation: .scanBlocks
+                    )
+
+                    let heightCount = lastScannedHeight - previousScannedHeight
+                    let seconds = scanFinishTime.timeIntervalSinceReferenceDate - scanStartTime.timeIntervalSinceReferenceDate
+                    LoggerProxy.debug("Scanned \(heightCount) blocks in \(seconds) seconds")
                 }
+
+                await Task.yield()
+            } while !Task.isCancelled && scannedNewBlocks && lastScannedHeight < targetScanHeight
+            if Task.isCancelled {
+                state = .stopped
+                LoggerProxy.debug("Warning: compactBlockBatchScanning cancelled")
             }
         } catch {
             LoggerProxy.debug("block scanning failed with error: \(String(describing: error))")
@@ -98,11 +93,12 @@ extension CompactBlockProcessor {
         }
     }
 
-    fileprivate func scanBatchSize(for range: CompactBlockRange, network: NetworkType) -> UInt32 {
+    fileprivate func scanBatchSize(startScanHeight height: BlockHeight, network: NetworkType) -> UInt32 {
+        assert(config.scanningBatchSize > 0, "ZcashSDK.DefaultScanningBatch must be larger than 0!")
         guard network == .mainnet else {
             return UInt32(config.scanningBatchSize)
         }
-        if range.lowerBound > 1_600_000 {
+        if height > 1_600_000 {
             return 5
         }